+++ tests/engine/inc/types.h
@@ -160,6 +160,13 @@ struct subtest_data {
 		_this->app->macs;                                              \
 	})
 
+#define list_certificates(this)                                                \
+	({                                                                     \
+		struct subtest_data *_this = (this);                           \
+		assert(_this->app);                                            \
+		_this->app->certificates;                                      \
+	})
+
 /**
  * struct tbuffer - Data of type buffer
  * @data: Data buffer