+++ tests/engine/inc/util_app.h
@@ -21,6 +21,7 @@
  * @ciphers:         Cipher to verify list
  * @signatures:      Signatures to verify list
  * @macs:            MACs to verify list
+ * @certificates:    Certificates to verify list
  * @threads:         Application threads list
  * @semaphores:      Semaphores list
  * @is_multithread:  Application is multithread
@@ -40,6 +41,7 @@ struct app_data {
 	struct llist *ciphers;
 	struct llist *signatures;
 	struct llist *macs;
+	struct llist *certificates;
 	struct llist *threads;
 	struct llist *semaphores;
 	int is_multithread;