+++ CHANGELOG.md
@@ -68,6 +68,7 @@ The failure is due to the storage manager which is already loaded and a new inst
 * Add tests to validate smw_device_get_lifecycle() and smw_device_set_lifecycle().
 * Add project option `TEE_TA_DESTDIR` to install TAs in a directory other than
   the system `\lib` directory.
+* Add more tests to validate device UUID and Attestation APIs.
 
 #### PKCS#11 Library - _version 2.5_